Update based on requirements.
Changed the three flower instances into a flower list, and added a for loop that moves, collides, and draws all flowers in the list.

The flower class now has xVel and yVel to control the movement speed of the flower (on each axis). Two additional methods was added, one to move the flower, and the other to check for collisions with the screen edges.